+++ source/sites/publicodes/conference/utils.tsx
@@ -1,0 +1,14 @@
+export const getRandomInt = (max) => Math.floor(Math.random() * Math.floor(max))
+
+export const stringToColour = function (str) {
+	var hash = 0
+	for (var i = 0; i < str.length; i++) {
+		hash = str.charCodeAt(i) + ((hash << 5) - hash)
+	}
+	var colour = '#'
+	for (var i = 0; i < 3; i++) {
+		var value = (hash >> (i * 8)) & 0xff
+		colour += ('00' + value.toString(16)).substr(-2)
+	}
+	return colour
+}
